Various fixes
* doc/cpio.texi: Document URLs of the mailing list archive and
subscription interface.
* src/copyin.c (try_existing_file): Change type of the last argument.
(copyin_file): Initialize existing_dir.
* src/dstring.c: Include xalloc.h
* src/filemode.c (mode_string): mode is unsigned int.

+++ src/filemode.c
@@ -191,7 +191,7 @@ setst (unsigned short bits, char *chars)
    `struct stat' is given as an argument.  */
 
 void
-mode_string (unsigned short mode, char *str)
+mode_string (unsigned int mode, char *str)
 {
   str[0] = ftypelet ((long) mode);
   rwx ((mode & 0700) << 0, &str[1]);
